Extra comments and tidy up
